ide: fix booting in mfg mode
In mfg mode, we want to load a rather large initramfs (~80MB). This
means wee need to tweak several variables in u-boot to make everything
run smoothly.

1- CONFIG_SYS_TEXT_BASE: move to 0x81f00000 which is 1MB lower than
the load address of the fit image (0x82000000). This will allow us to
grow the initramfs more easily. Also, 1MB is plenty for u-boot. We
also initially had a problem where the bootloader image (u-boot.img)
would overwrite the fit image in memory causing checksum errors.

2- CONFIG_SYS_BOOTM_LEN: increase the maximum amount of memory
supported by the 'bootm' command. Default is 16M.

The other changes in this commit are caused by the fact that we
stopped using the mx6_common.h include file.